requirements.txt
# BUTLER System Requirements
# Python 3.8+ required

# Core Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
werkzeug==3.0.1

# HTTP Client (for Ollama communication)
requests==2.31.0

# Document Processing
PyPDF2==3.0.1

# Async Support (optional, for advanced features)
aiohttp==3.9.1
asyncio==3.4.3

# Environment Management
python-dotenv==1.0.0

# Security and encryption
cryptography==41.0.3

# Utilities
numpy==1.24.3
python-dateutil==2.8.2

# Development (optional)
pytest==7.4.2
pytest-asyncio==0.21.1